Move remaining stuff out of legacy blueprint
it was just marcxml and event roombooking anyway
